+++ icekit/admin_tools/apps.py
@@ -1,7 +1,7 @@
 from django.apps import AppConfig
 
 
-class DashboardConfig(AppConfig):
+class AppConfig(AppConfig):
     name = '.'.join(__name__.split('.')[:-1])  # Portable
     label = 'icekit_dashboard'