+++ icekit/page_types/layout_page/admin.py
@@ -1,6 +1,6 @@
 from fluent_pages.integration.fluent_contents.admin import FluentContentsPageAdmin
 
-from icekit.admin_mixins import FluentLayoutsMixin, HeroMixinAdmin, \
+from icekit.admin_tools.mixins import FluentLayoutsMixin, HeroMixinAdmin, \
     ListableMixinAdmin
 from icekit.admin import ICEkitContentsAdmin