+++ docs/changelog.rst
@@ -6,6 +6,13 @@ In development
 
 -  Fleshed out Image model with more metadata and fields for rights/usage.
 
+-  Add ``DASHBOARD_SORTED_APPS`` setting to specify models that should appear
+   first.
+
+-  Refactor icekit.admin* and icekit.utils.admin.* to icekit.admin_tools.*
+   in preparation for admin consolidation and extension. Previous classes and
+   functions are deprecated.
+
 -  Add ``icekit.workflow`` application to associate, manage, and filter
    workflow state information like status and user-assigment for
    arbitrary models.
@@ -61,6 +68,11 @@ In development
 
 Breaking changes
 ~~~~~~~~~~~~~~~~
+
+-  ``DASHBOARD_FEATURED_APPS[]['models']`` should now be a list of tuples, not
+   a dict, so that ordering is preserved. Dict-based definitions still work, but
+   operations that rely on the value being a dict (such as ``update()``)
+   will fail.
 
 -  ``AbstractLayoutPage`` now includes ListableMixin and HeroMixin. All
    models which inherit from this will need a new migration.
@@ -121,7 +133,7 @@ Breaking changes
    ``icekit.abstract_models``.
 
 -  Import admin mixin ``FluentLayoutsMixin`` from
-   ``icekit.admin_mixins`` module instead of ``icekit.admin``.
+   ``icekit.admin_tools.mixins`` module instead of ``icekit.admin``.
 
 0.14.1 (2016-09-26)
 -------------------